@@ -1,21 +1,18 @@
 /* eslint-disable @typescript-eslint/ban-types */
-import React, { Component, ComponentClass, ComponentType } from 'react';
-import { ConsumerType, EqualityCheckFn, SelectorCallback } from './types';
+import React, { ComponentType } from 'react';
+import { EqualityCheckFn, SelectorCallback, UseSelectorHookDef } from './types';
 
-export function createHOCFunction<T>(Consumer: ConsumerType<T>) {
+export function createHOCFunction<T>(useSelector: UseSelectorHookDef<T>) {
   return function withSelector<R>(
     selector: SelectorCallback<T, R>,
     equalityFn?: EqualityCheckFn
-  ): <P extends object>(WrappedComponent: ComponentType<P>) => ComponentClass<P> {
-    return <P extends object>(WrappedComponent: ComponentType<P>) =>
-      class SelectorContextComponent extends Component<P> {
-        public render() {
-          return (
-            <Consumer<R> selector={selector} equalityFn={equalityFn}>
-              {value => (<WrappedComponent {...this.props} selectorValue={value} />)}
-            </Consumer>
-          );
-        }
-      };
+  ) {
+    return <P extends object>(WrappedComponent: ComponentType<P & {selectorValue: R}>) => {
+      const SelectorContextComponent: React.FC<P> = function SelectorContextComponent(props: P) {
+        const value = useSelector<R>(selector, equalityFn);
+        return <WrappedComponent {...props} selectorValue={value} />
+      }
+      return SelectorContextComponent;
+    };
   }
 }
